feat: Provide `autoQuery` configuration to disable debounced auto query
Soon we will expand the component to be a full form with a submit button. In this case always doing
queries after changing the input value might not be desired. Controlling the input value through a
separate state will allow us to retrieve its value beyond just a form event.