refactor: Build in checks if contexts are provided
Contexts must be provided and if they are not errors should happen. In order for the Suggestions
component to be used independendly of the searchbox or with a different popup library we should
allow floating UI context to not be defined (or used). Allowing the contexts to be null also allows
using the 'hooks' without the Context being available as the values are also returned from the
hooks. Setting the contexts to be potentially 'null' is also more aligned with what could happen
instead of providing dummy implementations that will silently not work as expected.